chore: refactor useLyric
