+++ src/hooks/useLyric.ts
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import i18n from 'i18next';
 
 import { reExtractSongName } from '@stores/appStore';
